refactor: rename @mui/x-virtualizer to @base-ui/virtualizer
Rename the package and move packages/x-virtualizer to packages/virtualizer,
following the folder = unscoped-name convention.
