+++ packages/virtualizer/README.md
@@ -1,3 +1,3 @@
-# x-virtualizer
+# @base-ui/virtualizer
 
 Virtualization engine
