+++ packages/virtualizer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@mui/x-virtualizer",
+  "name": "@base-ui/virtualizer",
   "version": "0.6.3",
   "author": "MUI Team",
   "description": "MUI virtualization library",
@@ -30,7 +30,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/mui/mui-x.git",
-    "directory": "packages/x-virtualizer"
+    "directory": "packages/virtualizer"
   },
   "dependencies": {
     "@babel/runtime": "catalog:",
