+++ packages/virtualizer/src/features/virtualization/anchor.ts
