+++ packages/virtualizer/src/models/colspan.ts
