+++ packages/virtualizer/src/models/core.ts
