+++ packages/x-data-grid/src/components/virtualization/GridVirtualScroller.tsx
@@ -4,7 +4,7 @@ import { styled } from '@mui/material/styles';
 import composeClasses from '@mui/utils/composeClasses';
 import { iOSMediaQuery } from '@mui/x-internals/platform';
 import clsx from 'clsx';
-import { LayoutDataGrid, Virtualization } from '@mui/x-virtualizer';
+import { LayoutDataGrid, Virtualization } from '@base-ui/virtualizer';
 import {
   gridHasBottomFillerSelector,
   gridHasScrollXSelector,
