+++ packages/x-data-grid/src/hooks/features/dimensions/gridDimensionsApi.ts
@@ -1,4 +1,4 @@
-import type { DimensionsState } from '@mui/x-virtualizer/models';
+import type { DimensionsState } from '@base-ui/virtualizer/models';
 
 export interface GridDimensions extends DimensionsState {
   /**
