+++ packages/x-data-grid/src/hooks/features/rows/gridRowsMetaState.ts
@@ -1,4 +1,4 @@
-import type { RowsMetaState } from '@mui/x-virtualizer/models';
+import type { RowsMetaState } from '@base-ui/virtualizer/models';
 
 /**
  * The grid rows total height and row positions.
