+++ packages/x-data-grid/src/hooks/features/virtualization/index.ts
@@ -1,4 +1,4 @@
-export { EMPTY_RENDER_CONTEXT } from '@mui/x-virtualizer';
+export { EMPTY_RENDER_CONTEXT } from '@base-ui/virtualizer';
 export * from './useGridVirtualization';
 export {
   gridVirtualizationSelector,
