+++ packages/x-data-grid/tsconfig.build.json
@@ -18,7 +18,7 @@
   },
   "references": [
     { "path": "../x-internals/tsconfig.build.json" },
-    { "path": "../x-virtualizer/tsconfig.build.json" }
+    { "path": "../virtualizer/tsconfig.build.json" }
   ],
   "include": ["src/**/*.ts*"],
   "exclude": ["src/**/*.spec.ts*", "src/**/*.test.ts*"]
