+++ packages/x-scheduler-premium/package.json
@@ -60,13 +60,13 @@
     "@babel/runtime": "catalog:",
     "@base-ui/react": "catalog:",
     "@base-ui/utils": "catalog:",
+    "@base-ui/virtualizer": "workspace:^",
     "@mui/utils": "catalog:",
     "@mui/x-internals": "workspace:^",
     "@mui/x-license": "workspace:^",
     "@mui/x-scheduler": "workspace:^",
     "@mui/x-scheduler-internals": "workspace:^",
     "@mui/x-scheduler-internals-premium": "workspace:^",
-    "@mui/x-virtualizer": "workspace:^",
     "clsx": "catalog:",
     "prop-types": "catalog:"
   },
