+++ packages/x-scheduler-premium/tsconfig.build.json
@@ -17,7 +17,7 @@
     { "path": "../x-scheduler-internals-premium/tsconfig.build.json" },
     { "path": "../x-license/tsconfig.build.json" },
     { "path": "../x-internals/tsconfig.build.json" },
-    { "path": "../x-virtualizer/tsconfig.build.json" }
+    { "path": "../virtualizer/tsconfig.build.json" }
   ],
   "include": ["src/**/*.ts*"],
   "exclude": ["src/**/*.spec.ts*", "src/**/*.test.ts*", "src/tests/**/*"]
