+++ packages/x-tree-view-pro/package.json
@@ -39,11 +39,11 @@
   "dependencies": {
     "@babel/runtime": "catalog:",
     "@base-ui/utils": "catalog:",
+    "@base-ui/virtualizer": "workspace:*",
     "@mui/utils": "catalog:",
     "@mui/x-internals": "workspace:^",
     "@mui/x-license": "workspace:^",
     "@mui/x-tree-view": "workspace:^",
-    "@mui/x-virtualizer": "workspace:*",
     "@types/react-transition-group": "catalog:",
     "clsx": "catalog:",
     "prop-types": "catalog:",
